Fix error handling in integration test script.
- Switch destructor to TearDown for the TestIntegration class as the
  tmp_files variable is only in the scope of the individual test.
- Explicitly print the error message that is returned when srun fails.
  This was being swallowed by the unit test framework.
- Add back the exception checking logic in the file delete routines.
  This was erroneously removed before.
- Set global var to ensure files are preserved on a failed test. exc_info()
  is unreliable.

+++ test_integration/geopm_io.py
@@ -37,6 +37,8 @@ import re
 import pandas
 import fnmatch
 
+keep_files=False
+
 class AppOutput(object):
     def __init__(self, report_base, trace_base=None):
         self._reports = {}
@@ -57,7 +59,7 @@ class AppOutput(object):
                             [Trace(tp) for tp in trace_paths]}
 
     def __del__(self):
-        if os.getenv('GEOPM_KEEP_FILES') is None:
+        if keep_files == False and os.getenv('GEOPM_KEEP_FILES') is None:
             for ff in self._all_paths:
                 try:
                     os.remove(ff)
